refactor(app): separate routes and pages
