+++ frontend/src/app/routeTree.gen.ts
@@ -10,74 +10,74 @@
 
 // Import Routes
 
-import { Route as rootRoute } from './../pages/__root'
-import { Route as TeamsImport } from './../pages/teams'
-import { Route as SearchImport } from './../pages/search'
-import { Route as IndexImport } from './../pages/index'
-import { Route as SearchIndexImport } from './../pages/search.index'
-import { Route as SearchTeamsImport } from './../pages/search.teams'
-import { Route as SearchMembersImport } from './../pages/search.members'
+import { Route as rootRoute } from './routes/__root';
+import { Route as TeamsImport } from './routes/teams';
+import { Route as SearchImport } from './routes/search';
+import { Route as IndexImport } from './routes/index';
+import { Route as SearchIndexImport } from './routes/search.index';
+import { Route as SearchTeamsImport } from './routes/search.teams';
+import { Route as SearchMembersImport } from './routes/search.members';
 
 // Create/Update Routes
 
 const TeamsRoute = TeamsImport.update({
   path: '/teams',
   getParentRoute: () => rootRoute,
-} as any)
+} as any);
 
 const SearchRoute = SearchImport.update({
   path: '/search',
   getParentRoute: () => rootRoute,
-} as any)
+} as any);
 
 const IndexRoute = IndexImport.update({
   path: '/',
   getParentRoute: () => rootRoute,
-} as any)
+} as any);
 
 const SearchIndexRoute = SearchIndexImport.update({
   path: '/',
   getParentRoute: () => SearchRoute,
-} as any)
+} as any);
 
 const SearchTeamsRoute = SearchTeamsImport.update({
   path: '/teams',
   getParentRoute: () => SearchRoute,
-} as any)
+} as any);
 
 const SearchMembersRoute = SearchMembersImport.update({
   path: '/members',
   getParentRoute: () => SearchRoute,
-} as any)
+} as any);
 
 // Populate the FileRoutesByPath interface
 
 declare module '@tanstack/react-router' {
   interface FileRoutesByPath {
     '/': {
-      preLoaderRoute: typeof IndexImport
-      parentRoute: typeof rootRoute
-    }
+      preLoaderRoute: typeof IndexImport;
+      parentRoute: typeof rootRoute;
+    };
     '/search': {
-      preLoaderRoute: typeof SearchImport
-      parentRoute: typeof rootRoute
-    }
+      preLoaderRoute: typeof SearchImport;
+      parentRoute: typeof rootRoute;
+    };
     '/teams': {
-      preLoaderRoute: typeof TeamsImport
-      parentRoute: typeof rootRoute
-    }
+      preLoaderRoute: typeof TeamsImport;
+      parentRoute: typeof rootRoute;
+    };
     '/search/members': {
-      preLoaderRoute: typeof SearchMembersImport
-      parentRoute: typeof SearchImport
-    }
+      preLoaderRoute: typeof SearchMembersImport;
+      parentRoute: typeof SearchImport;
+    };
     '/search/teams': {
-      preLoaderRoute: typeof SearchTeamsImport
-      parentRoute: typeof SearchImport
-    }
+      preLoaderRoute: typeof SearchTeamsImport;
+      parentRoute: typeof SearchImport;
+    };
     '/search/': {
-      preLoaderRoute: typeof SearchIndexImport
-      parentRoute: typeof SearchImport
-    }
+      preLoaderRoute: typeof SearchIndexImport;
+      parentRoute: typeof SearchImport;
+    };
   }
 }
 
@@ -85,12 +85,8 @@ declare module '@tanstack/react-router' {
 
 export const routeTree = rootRoute.addChildren([
   IndexRoute,
-  SearchRoute.addChildren([
-    SearchMembersRoute,
-    SearchTeamsRoute,
-    SearchIndexRoute,
-  ]),
+  SearchRoute.addChildren([SearchMembersRoute, SearchTeamsRoute, SearchIndexRoute]),
   TeamsRoute,
-])
+]);
 
 /* prettier-ignore-end */
