+++ frontend/src/app/routes/__root.tsx
@@ -1,9 +1,5 @@
 import { createRootRoute, Outlet } from '@tanstack/react-router';
 
 export const Route = createRootRoute({
-  component: () => (
-    <>
-      <Outlet />
-    </>
-  ),
+  component: () => <Outlet />,
 });
