+++ frontend/src/app/routes/search.index.tsx
