+++ frontend/src/app/routes/search.teams.tsx
@@ -1,0 +1,6 @@
+import { SearchTeamPage } from '@pages/teams';
+import { createFileRoute } from '@tanstack/react-router';
+
+export const Route = createFileRoute('/search/teams')({
+  component: () => <SearchTeamPage />,
+});
