+++ frontend/src/app/routes/search.tsx
@@ -1,0 +1,6 @@
+import { SearchPage } from '@pages/search';
+import { createFileRoute } from '@tanstack/react-router';
+
+export const Route = createFileRoute('/search')({
+  component: () => <SearchPage />,
+});
