+++ frontend/src/app/routes/teams.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { Page, Spinner } from '@ui';
+import { Page, Spinner } from '@shared/ui';
 
 export const Route = createFileRoute('/teams')({
   component: () => (
