+++ frontend/src/pages/search/ui.tsx
@@ -1,13 +1,9 @@
 import { css } from '@style/css';
-import { Outlet, createFileRoute, useNavigate, useLocation } from '@tanstack/react-router';
-import { Page, PageTitle, TabsPanel } from '@ui';
+import { Outlet, useLocation, useNavigate } from '@tanstack/react-router';
+import { Page, PageTitle, TabsPanel } from '@shared/ui';
 import { NavMenu } from '@widgets';
 
-export const Route = createFileRoute('/search')({
-  component: () => <SearchPage />,
-});
-
-function SearchPage() {
+export function SearchPage() {
   const navigate = useNavigate();
   const tabId = useLocation({ select: ({ pathname }) => pathname.split('/').at(-1) });
