+++ frontend/src/pages/teams/index.ts
@@ -1,0 +1,1 @@
+export { SearchTeamPage } from './ui';
