+++ frontend/src/shared/api/index.ts
@@ -1,1 +1,5 @@
+import { Api } from './generated';
+
 export * from './generated';
+
+export const api = new Api();
