+++ frontend/tsr.config.json
@@ -1,4 +1,4 @@
 {
-  "routesDirectory": "./src/pages",
+  "routesDirectory": "./src/app/routes",
   "generatedRouteTree": "./src/app/routeTree.gen.ts"
 }
